Initial Assignment 4 Work: Code is retrieving forecast data correctly.
