+++ app/build.gradle.kts
@@ -90,4 +90,9 @@ dependencies {
     implementation(libs.kotlinx.coroutines.core)
     implementation(libs.kotlinx.coroutines.android)
     implementation(libs.androidx.navigation.compose)
+    dependencies {
+    // Navigation Compose
+    implementation("androidx.navigation:navigation-compose:2.7.6")
+
+    }
 }
